status.rsをinfer/下から引っ張り出す
