+++ crates/voicevox_core/src/infer.rs
@@ -1,7 +1,7 @@
 pub(crate) mod domains;
 mod model_file;
 pub(crate) mod runtimes;
-pub(crate) mod status;
+pub(crate) mod session_set;
 
 use std::{
     borrow::Cow, collections::BTreeSet, convert::Infallible, fmt::Debug, marker::PhantomData,
