+++ crates/voicevox_core/src/status.rs
@@ -1,27 +1,24 @@
-use std::{
-    any, collections::HashMap, convert::Infallible, fmt::Display, marker::PhantomData, sync::Arc,
-};
+use std::{any, convert::Infallible, marker::PhantomData};
 
-use anyhow::bail;
 use derive_new::new;
 use educe::Educe;
-use enum_map::{Enum as _, EnumMap};
+use enum_map::EnumMap;
 use indexmap::IndexMap;
 use itertools::{iproduct, Itertools as _};
 
 use crate::{
     error::{ErrorRepr, LoadModelError, LoadModelErrorKind, LoadModelResult},
+    infer::{
+        session_set::{SessionCell, SessionSet},
+        ForAllInferenceDomain, InferenceDomain, InferenceDomainGroup, InferenceDomainMap as _,
+        InferenceDomainMapValueFunction, InferenceDomainMapValuePredicate,
+        InferenceDomainMapValueProjection, InferenceDomainMapValueTryFunction,
+        InferenceInputSignature, InferenceRuntime, InferenceSessionOptions, InferenceSignature,
+    },
     manifest::{ModelInnerId, StyleIdToModelInnerId},
     metas::{self, SpeakerMeta, StyleId, StyleMeta, VoiceModelMeta},
     voice_model::{ModelBytesByInferenceDomain, VoiceModelHeader, VoiceModelId},
     Result, StyleType,
-};
-
-use super::{
-    model_file, ForAllInferenceDomain, InferenceDomain, InferenceDomainGroup,
-    InferenceDomainMap as _, InferenceDomainMapValueFunction, InferenceDomainMapValuePredicate,
-    InferenceDomainMapValueProjection, InferenceDomainMapValueTryFunction, InferenceInputSignature,
-    InferenceOperation, InferenceRuntime, InferenceSessionOptions, InferenceSignature, ParamInfo,
 };
 
 pub(crate) struct Status<R: InferenceRuntime, G: InferenceDomainGroup> {
@@ -368,93 +365,6 @@ impl<R: InferenceRuntime, G: InferenceDomainGroup> LoadedModels<R, G> {
     }
 }
 
-struct SessionSet<R: InferenceRuntime, D: InferenceDomain>(
-    EnumMap<D::Operation, Arc<std::sync::Mutex<R::Session>>>,
-);
-
-impl<R: InferenceRuntime, D: InferenceDomain> SessionSet<R, D> {
-    fn new(
-        model_bytes: &EnumMap<D::Operation, Vec<u8>>,
-        options: &EnumMap<D::Operation, InferenceSessionOptions>,
-    ) -> anyhow::Result<Self> {
-        let mut sessions = model_bytes
-            .iter()
-            .map(|(op, model_bytes)| {
-                let (expected_input_param_infos, expected_output_param_infos) =
-                    <D::Operation as InferenceOperation>::PARAM_INFOS[op];
-
-                let (sess, actual_input_param_infos, actual_output_param_infos) =
-                    R::new_session(|| model_file::decrypt(model_bytes), options[op])?;
-
-                check_param_infos(expected_input_param_infos, &actual_input_param_infos)?;
-                check_param_infos(expected_output_param_infos, &actual_output_param_infos)?;
-
-                Ok((op.into_usize(), std::sync::Mutex::new(sess).into()))
-            })
-            .collect::<anyhow::Result<HashMap<_, _>>>()?;
-
-        return Ok(Self(EnumMap::<D::Operation, _>::from_fn(|k| {
-            sessions.remove(&k.into_usize()).expect("should exist")
-        })));
-
-        fn check_param_infos<D: PartialEq + Display>(
-            expected: &[ParamInfo<D>],
-            actual: &[ParamInfo<D>],
-        ) -> anyhow::Result<()> {
-            if !(expected.len() == actual.len()
-                && itertools::zip_eq(expected, actual)
-                    .all(|(expected, actual)| expected.accepts(actual)))
-            {
-                let expected = display_param_infos(expected);
-                let actual = display_param_infos(actual);
-                bail!("expected {{{expected}}}, got {{{actual}}}")
-            }
-            Ok(())
-        }
-
-        fn display_param_infos(infos: &[ParamInfo<impl Display>]) -> impl Display {
-            infos
-                .iter()
-                .map(|ParamInfo { name, dt, ndim }| {
-                    let brackets = match *ndim {
-                        Some(ndim) => "[]".repeat(ndim),
-                        None => "[]...".to_owned(),
-                    };
-                    format!("{name}: {dt}{brackets}")
-                })
-                .join(", ")
-        }
-    }
-}
-
-impl<R: InferenceRuntime, D: InferenceDomain> SessionSet<R, D> {
-    fn get<I>(&self) -> SessionCell<R, I>
-    where
-        I: InferenceInputSignature,
-        I::Signature: InferenceSignature<Domain = D>,
-    {
-        SessionCell {
-            inner: self.0[I::Signature::OPERATION].clone(),
-            marker: PhantomData,
-        }
-    }
-}
-
-struct SessionCell<R: InferenceRuntime, I> {
-    inner: Arc<std::sync::Mutex<R::Session>>,
-    marker: PhantomData<fn(I)>,
-}
-
-impl<R: InferenceRuntime, I: InferenceInputSignature> SessionCell<R, I> {
-    fn run(self, input: I) -> crate::Result<<I::Signature as InferenceSignature>::Output> {
-        let inner = &mut self.inner.lock().unwrap();
-        let ctx = input.make_run_context::<R>(inner);
-        R::run(ctx)
-            .and_then(TryInto::try_into)
-            .map_err(|e| ErrorRepr::InferenceFailed(e).into())
-    }
-}
-
 pub(crate) enum SessionOptionsByDomain {}
 
 impl InferenceDomainMapValueProjection for SessionOptionsByDomain {
@@ -494,13 +404,16 @@ mod tests {
     use rstest::rstest;
 
     use crate::{
-        infer::domains::{InferenceDomainGroupImpl, InferenceDomainMapImpl, TalkOperation},
+        infer::{
+            domains::{InferenceDomainGroupImpl, InferenceDomainMapImpl, TalkOperation},
+            InferenceSessionOptions,
+        },
         macros::tests::assert_debug_fmt_eq,
         synthesizer::InferenceRuntimeImpl,
         test_util::open_default_vvm_file,
     };
 
-    use super::{super::InferenceSessionOptions, Status};
+    use super::Status;
 
     #[rstest]
     #[case(true, 0)]
