+++ crates/voicevox_core/src/synthesizer.rs
@@ -88,9 +88,9 @@ pub(crate) mod blocking {
                 PredictDurationInput, PredictDurationOutput, PredictIntonationInput,
                 PredictIntonationOutput, TalkDomain, TalkOperation,
             },
-            status::Status,
             InferenceSessionOptions,
         },
+        status::Status,
         text_analyzer::{KanaAnalyzer, OpenJTalkAnalyzer, TextAnalyzer},
         AccentPhraseModel, AudioQueryModel, FullcontextExtractor, Result, StyleId,
         SupportedDevices, SynthesisOptions, VoiceModelId, VoiceModelMeta,
